# Optimize group_norm_op