… caffe2/operators/group_norm_op.cu