Summary: Optimize group_norm_op

Differential Revision: D14419908