@BIT-silence 2019-03-12
Some benchmarks.

For N = 128, C = 256, H = 112, W = 112, group = 32 on GPU:
NCHW backward: 56.75ms -> 38.45ms
NHWC backward: 532.87ms -> 84.39ms

For N = 64, C = 256, H = 56, W = 56, group = 32 on CPU:
NCHW backward: 768.14ms -> 97.17ms
NHWC forward: 623.72 -> 92.20ms
NHWC backward: 764.38 -> 121.68ms